refactor(): extract utils
